requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncio==3.4.3
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
decorator==5.1.1
deepdiff==5.8.1
fastapi==0.79.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
motor==3.0.0
ordered-set==4.1.0
pydantic==1.9.2
pymongo==4.2.0
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tornado==6.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.14.0
validators==0.18.2
